requirements.txt
# Core framework and UI
streamlit==1.45.1
fastapi==0.104.1
uvicorn==0.24.0
tqdm==4.66.2
python-dotenv==1.0.0

# LangChain ecosystem
langchain==0.3.25
langchain-core==0.3.60
langchain-community==0.3.24
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.3.42

# Vector store + Embeddings
faiss-cpu==1.11.0
sentence-transformers==4.1.0
huggingface-hub==0.31.2
transformers==4.51.3
torch==2.3.0

# Data + utilities
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.5.2
requests==2.31.0
pydantic==2.12.4
pydantic-settings==2.11.0

# Optional helpers
python-multipart==0.0.6
orjson==3.10.18
pyyaml==6.0.1
